Improved the hierarchy-input UX, by making the prompt icon the icon of the recently selected item
Improved the hierarchy-input UX, by making it take up available space in forms

Improved robustness of hierarchy-input

Added support for custom 'value properties' for multi-flex-form. This means that input types could be embedded that use different properties, instead of 'value'
to get and set the value of the user's input. Take for example input type 'date', could use 'valueAsNumber' instead of just 'value'

Improved usability of GenericListings widget, by allowing arguments to be passed in the constructor

Improved the design of the uniqueFileUpload widget

Improved the typing of the AlarmObject class

Created a new widget called prompt-input that allows for easily creating inputs that require the user to tap a button, and follow a popup

Improved robustness of the google-location-input widget

Removed unnecessary logging from the file-explorer widget because it is now stable.