.
Made the waitTillDOMAttached() method more efficient

Added a feature which widgets can take advantage. The class hc-widget-ready, is automatically added to the html of a widget the first time it is appended to the DOM.

Improved design of the back-forth widget

Improved usability of the action-button widget

Made slide-container widget more resource-efficient, by limiting the time it takes to wait for the slide process to complete.