.
Improved the loading UI of all widgets.

Added a utility called loading-action, which permits widgets to perform actions while loading, in a manner that allows the user to retry an action, if it fails. It is available via the `blockWithAction()` method

Improved code quality of the widget-image utility, by properly naming CSS classes.

Improved the auto-css import feature.

Improved the popup-menu widget, by improving the logic for the detection of outside clicks.

Corrected minor bug with the watch-to-css utility.

Improved the default layout of images imported with the widget-image utility.